uitoa

+++ ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: vgroux <vgroux@student.42lausanne.ch>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/18 15:55:55 by vgroux            #+#    #+#             */
-/*   Updated: 2022/10/21 14:44:39 by vgroux           ###   ########.fr       */
+/*   Updated: 2022/10/21 16:07:06 by vgroux           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,7 +30,7 @@ int	ft_printf_conv(char format, va_list ap)
 	else if (format == 'X')
 		len = ft_printf_nbr_base(va_arg(ap, unsigned int), "0123456789ABCDEF");
 	else if (format == 'p')
-		len = ft_printf_ptr(va_arg(ap, uintptr_t));
+		len = ft_printf_ptr(va_arg(ap, unsigned long));
 	else if (format == 'u')
 		len = ft_printf_nbr_base(va_arg(ap, unsigned long), "0123456789");
 	return (len);
